DOPE-170: added subqueries
Made the select clause extend TypeExpression<ArrayType<ValidType>> that it can be used like an Array Expression. The SelectRaw also has a generic for the ArrayType that comparisons or inArray Expression can be created using subqueries.
All Resolvable that use a TypeExpression<ValidType> needs to check if the TypeExpression is a select to add the round brackets to only generate valid syntax.